browser/extensions/formautofill/test/mochitest/test_on_address_submission.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofill submit</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="formautofill_common.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Form autofill test: check if address is saved/updated correctly

<script>
/* import-globals-from ../../../../../testing/mochitest/tests/SimpleTest/SpawnTask.js */
/* import-globals-from ../../../../../toolkit/components/satchel/test/satchel_common.js */
/* import-globals-from formautofill_common.js */

"use strict";

let TEST_ADDRESSES = [{
  organization: "Sesame Street",
  "street-address": "123 Sesame Street.",
  tel: "1-345-345-3456",
}, {
  organization: "Mozilla",
  "street-address": "331 E. Evelyn Avenue",
  tel: "1-650-903-0800",
}];

// Autofill the address from dropdown menu.
add_task(async function check_storage_after_form_submitted() {
  // We already verified the first time use case in browser test
  await SpecialPowers.pushPrefEnv({
    set: [["extensions.formautofill.firstTimeUse", false]],
  });

  for (let key in TEST_ADDRESSES[0]) {
    await setInput("#" + key, TEST_ADDRESSES[0][key]);
  }

  clickOnElement("input[type=submit]");

  let expectedAddresses = TEST_ADDRESSES.slice(0, 1);
  await onAddressChanged("add");
  let matching = await checkAddresses(expectedAddresses);
  ok(matching, "Address saved as expected");
});

</script>

<div>

  <form onsubmit="return false">
    <p>This is a basic form for submitting test.</p>
    <p><label>organization: <input id="organization" name="organization" autocomplete="organization" type="text"></label></p>
    <p><label>streetAddress: <input id="street-address" name="street-address" autocomplete="street-address" type="text"></label></p>
    <p><label>tel: <input id="tel" name="tel" autocomplete="tel" type="text"></label></p>
    <p><label>country: <input id="country" name="country" autocomplete="country" type="text"></label></p>
    <p><input type="submit"></p>
  </form>

</div>
</body>
</html>
